Add files via upload
This is the script to pull data from the google sheet budget and places it in a csv. To, run you need a google drive + sheets api key + client set up. TODO: Be able to pull the hyperlinks to the articles from the sheet (right now only gets the link text) -> might need to write a custom script.